[SCB-2687]fix checkstyle and pmd problems
